Major refactor
Since the previous code was quite messy, this is a first step at cleaning the whole thing up.

The main goal was to strictly seperate between the game logic and the rendering process. Another was to loosen up the coupling between different modules and to keep the core simple. This should be especially beneficial later on for generating a gametree.

So far this refactor breaks the current bot and the rendering is currently highly inefficient, since everything gets reinitialized each time the scene gets rendered.